Added new control ( string/labels ).  String is of the form eg. "0,0,1,1,2,2," such that each successive label is assigned to each succesive processed sample.  If there are more samples than labels in the control, the labels of string/labels are repeated until all samples have been processed.
If string/labels == "", then the label assigned is the one held in the control natural/label.

This addition gets around the problem of not being able to assign different samples in a single
input realvec with different labels, such as after a PCA marsystem, when samples of multiple
classes are output at the same time (ie. in the same output realvec).




git-svn-id: svn+ssh://svn.code.sf.net/p/marsyas/code/trunk@29 0e98fd9d-079e-457f-9c5c-d482923f975d