Fix int->mrs_natural in tutorial.
See, here's a perfect example of a good granular commit.  The log
message is to the point, and it takes 10 seconds to scroll through the
actual diff to make sure that this is all I'm doing.  If I waited
another hour and make a commit that rolled this fix with a tons of other
things, it would be way harder to read the patch.


git-svn-id: svn+ssh://svn.code.sf.net/p/marsyas/code/trunk@2028 0e98fd9d-079e-457f-9c5c-d482923f975d