Finally finished the Spectrum2Chroma MarSystem. About 10 lines of code in MATLAB explode into a lot of C++ nested for loops :-\
Btw, PowerSpectrum does not output the fft N/2+1 unique bins magnitudes, but only N/2 values. This may be a little details not that important, but for the sake of DSP correctness, it should output N/2+1 magnitude values (where N is the size/order of the fft spectrum vector at its input). 

git-svn-id: svn+ssh://svn.code.sf.net/p/marsyas/code/trunk@2151 0e98fd9d-079e-457f-9c5c-d482923f975d