Made OggFileSource.h not include vorbis headers.
By changing the OggFileSource class to use untyped pointers,
the OggFileSource header no longer includes the vorbis headers,
thus requiring all files that include it to do so as well. This
also means that the ifdefs used to enable vorbis support in the
header file are no longer needed there either.

The vorbis headers and the relevant ifdefs are still needed
in OggFileSource.cpp, but those are only important when compiling
OggFileSource.cpp. No other file needs to have the vorbis headers
included to compile, and no other compilation needs to worry about
the MARSYAS_VORBIS preprocessor symbol. This should speed up and
decrease the memory requirements for the compilation of any source
file that needs OggFileSource.h (which is also included in
SoundFileSource.h and MarSystemManager.h )

I may do these optimizations for other classes soon.


git-svn-id: svn+ssh://svn.code.sf.net/p/marsyas/code/trunk@2580 0e98fd9d-079e-457f-9c5c-d482923f975d