Fixed distcheck problem with regressionTests.
Apparently, when creating the dist tarball, automake puts the headers for regressionChecks in a subfolder src/apps/regressionChecks/regressionChecks instead ov where they are in the SVN tree (one level higher), and as a result the compile fails. This simple fix adds the new directory to the list of include paths.


git-svn-id: svn+ssh://svn.code.sf.net/p/marsyas/code/trunk@2588 0e98fd9d-079e-457f-9c5c-d482923f975d