Fixed distclean problem with swig bindings.
It seems that when performing the distcheck, for a EXTRA_DIST directory, it creates Makefiles to recursively add the contents, and if a makefile exists already it assumes it was created by automake and will do everything correctly. This is a bad assumption to make, because in the swig directory there were already non-automake Makefiles, and thus caused distcheck to fail when certain commands were not supported.

To fix the problems, I renamed the Makefiles in a given swig directory to swin/LANG/Makefile.LANG, and created a swig/LANG/Makefile.am which automatically forwards 'all', 'clean', and 'install' targets to Makefile.LANG. Also I set up EXTRA_DIST and DISTCLEANFILES in these Makefile.am files so that 'make dist' and make 'distclean' work properly.

The advantage of this operation is that instead of doing a blind, wide, EXTRA_DIST that grabs all of swig, only the desired files are made part of the distribution. The disadvantage is that several more files have been added to junk the project up.

With this and the following (generated files) commit, the distchecks work on my linux machine.


git-svn-id: svn+ssh://svn.code.sf.net/p/marsyas/code/trunk@2590 0e98fd9d-079e-457f-9c5c-d482923f975d