mkcollection now accepts any extension.
If the user gives a filename with an extension, mkcollection
doesn't change the collection name.  If not, it adds a .mf to the
collection name.

This also fixes a crash with collection names less than 4
characters; this is particularly relevant to the GTZAN genre
database, which includes a "pop" directory (which triggered a
crash when using mkcollection).


git-svn-id: svn+ssh://svn.code.sf.net/p/marsyas/code/trunk@4545 0e98fd9d-079e-457f-9c5c-d482923f975d