Add -os --onlyStable to bextract.
(theoretically) ignores "unstable" outputs, such as the very first
instance of Flux (where the previous window is 0 and is therefore
probably not a reliable statistic about the audio).

Not all MarSystems have
  addToStabilizingDelay_
set correctly, notably Memory.  Some of that work will be done
this weekend, when I finally start "certifying" MarSystems.


git-svn-id: svn+ssh://svn.code.sf.net/p/marsyas/code/trunk@4559 0e98fd9d-079e-457f-9c5c-d482923f975d