add wreckBeach binary since mudbox takes too long to compile.
wreckBeach is a container executable for various simple
functions/applications that typically test drive a single
MarSystem or type of processing.

It is a place for the often-times ugly, unobscured source code.
No pretence of polite society is here!  Be bold, be adventurous.
Let it all hang out.  Have no shame!



git-svn-id: svn+ssh://svn.code.sf.net/p/marsyas/code/trunk@4583 0e98fd9d-079e-457f-9c5c-d482923f975d