correlation coefficient for regressions.
Wow this code is sucky.  I think I'm finally getting the hang of
academic programming!  I'm proud of myself.  :-)

In other news, I've signed up for the Stanford AI and machine
learning classes.  Partly because all my mcahine learning has come
in bits and pieces, but also partly because I think it's an
awesome experiment in education.  A class with over 100,000
students?!  using machine learning to teach and manage the vast
quantitiy of participants?!?!  Sign me up!  This is totally
uber-cool cyperpunk come to life!


git-svn-id: svn+ssh://svn.code.sf.net/p/marsyas/code/trunk@4652 0e98fd9d-079e-457f-9c5c-d482923f975d