Apply simple optimizations to Yin
Passes unit tests.  This marsystem was chewing up almost 40% of
Vivi's running time!

I love valgrind.  memcheck is great, but I've now discovered
cachegrind and callgrind.  I was really puzzled as to why a
10-second piece of audio took 0.34 seconds to generate in
artifastring but 15 seconds in Vivi -- I mean, sure the marsyas
feature extraction will take time, but surely not that much time.
And I was right!



git-svn-id: svn+ssh://svn.code.sf.net/p/marsyas/code/trunk@4679 0e98fd9d-079e-457f-9c5c-d482923f975d