Yin: use mrs_real* for interim use, and hard-code 1 channel.
This MarSystem was already hard-coded for only 1 channel, but it
used the very expensive realvec(r, c) access.  Changing this to
realvec(c) improved matters, and then changing to array[c]
improved things again.

Long-term, it would be good to change Yin to use an FFT for the
autocorrelation (as is done in AutoCorrelate), but that's too much
of a finnicky job for me to tackle until I've finished writing my
thesis.



git-svn-id: svn+ssh://svn.code.sf.net/p/marsyas/code/trunk@4681 0e98fd9d-079e-457f-9c5c-d482923f975d