add support for MARSYAS_DATADIR
If you're too lazy to check the docs, here's the blurb:
Collections support the environment variable MARSYAS_DATADIR. This
allows the use of .mf files shared between users (i.e. for a large
dataset of audio). For example, the above collection could be
rewritten as:

     MARSYAS_DATADIR/reggae/foo.wav \t music
     MARSYAS_DATADIR/reggae/bar.wav \t speech

provided that the user configures the environment variable
appropriately. For example, using bash on Linux or MacOS X, users
on three different machines may set up the variable as:

     export MARSYAS_DATADIR=/home/gtzan/data/sound/
     export MARSYAS_DATADIR=/Users/gtzan/data/sound/
     export MARSYAS_DATADIR=/home/gperciva/media/marsyas-data/


git-svn-id: svn+ssh://svn.code.sf.net/p/marsyas/code/trunk@4988 0e98fd9d-079e-457f-9c5c-d482923f975d