expand MRSERR() and MRSWARN() macros unconditionally
If an error or a warning is called for, something abnormal has happened,
so performance is not of importance. Hence, we can avoid the issue of
dependence of these macros on preprocessor definitions in header files.

On the other hand, MRSMSG and MRSDIAG which are often called in
performance-critical sections, are never called in a header file,
so dependence on preprocessor definitions is not an issue there.

git-svn-id: svn+ssh://svn.code.sf.net/p/marsyas/code/trunk@5061 0e98fd9d-079e-457f-9c5c-d482923f975d