don't pollute svn with sfplugin- tests.
Our testing framework is extremely fragile with respect to
directories.  :(   granted, bextract's directory handling doesn't
precisely help, in that it needs to be run from the same directory
that the mf file is in (because filenames within collections don't
inherit the directory prefix), plus the weirdness with outputDir
vs. workingDir.  :(

Unfortunately there's no quick fix to these problems.


git-svn-id: svn+ssh://svn.code.sf.net/p/marsyas/code/trunk@5112 0e98fd9d-079e-457f-9c5c-d482923f975d