potential fix for multi-file Timeline
My confidence is not outstanding because this is pretty messy.
TimelineLabeler's ctrl_advance is linked to CollectionFileSource's
advance.  This means that code in TimelineLabeler like:
  ctrl_advance->setValue(1);
  cout<<"advance: "<<ctrl_advance->to<mrs_natural>()<<endl;
produces
  advance: 0

This is because CFS::myUpdate handles the advance==1 case and then
sets advance to 0.  That meant that TimelineLabel::myUpdate never
got called when setting advance=1.  I extracted the "load a new
region" code into a separate function, and call that explicitly if
a new myAdvance variable is set.

Things get really confused when there's linked controls, both of
which have setctrlState(true), but one ::myUpdate() sets the
control with NOUPDATE.  Wait, delete the last two parts.  Things
get confused when there's linked controls, period.



git-svn-id: svn+ssh://svn.code.sf.net/p/marsyas/code/trunk@5120 0e98fd9d-079e-457f-9c5c-d482923f975d