Verify tempo reference up to bhisto
Hey kids, pop quiz: is multiplication commutative?

  double period = (int) (period + 0.5); // round to int
  int phase = 57;
  int b = 2;

  int index = (int) (phase - b*2*period);
  // this is equivalent to the upper line, right?
  // int index = (int) (phase - 2*b*period);

No, of course multiplcation isn't commutative!  Because period
isn't actually an int; it's still a double.  Using a double in
that multiplication induces different rounding behaviour, meaning
that if you change the order to make it easier to see the 1* 2* 3*
progression, you'll get different tempo strength results!  But not
a lot different.  No, a very different tempo result would be easy
to track down.  Instead, it's something like 0.1% of analysis
frames for which the rounding results in a different index being
calculated.

The most fun part?  IMO we should keep period as a double
*anyway*, since rounding it off to the nearest int serves no
purpose other than making the index calculations less accurate.
But the first step of a secondary implementation is to establish
bug-for-bug compatibility, so that'll have to wait for another
day...



git-svn-id: svn+ssh://svn.code.sf.net/p/marsyas/code/trunk@5132 0e98fd9d-079e-457f-9c5c-d482923f975d