continue file separation into src/marsyas/core and src/marsyas/marsystems
"src/marsyas/core" now contains all classes essentially related to
MarSystem.

There is a large amount of files left in the "src/marsyas" directory
which contain generally useful code, but not essentially related to
MarSystem, so I'm not very keen on including those in the "core" subdir.
Some of those are also used outside of the "src/marsyas" directory, so
we could consider moving them hierarchically higher.

git-svn-id: svn+ssh://svn.code.sf.net/p/marsyas/code/trunk@5189 0e98fd9d-079e-457f-9c5c-d482923f975d