unit_tests: fix building with static marsyas library
There appears to be no need for linking directly to the output path
of marsyas static library, at least on Linux. Moreover, that disables
the CMake functionality for transitive linking. So rather link to
the "marsyas" target itself.

Let's see if this builds on Windows and OS X as well...