scripting: destroy operation processor together with controlled MarSystem
This calls for a way to attach MarSystems to other MarSystems
so that the attached ones are deleted with those that they attach to.

In future, we could think of tighter integration of control expression
processing with the MarControl system itself, so as to avoid using
a specialized MarSystem for that purpose.